refactor: move ios/android utils into separate modules
