+++ packages/core/ui/transition/shared-transition-helper.ios.ts
@@ -3,7 +3,7 @@ import { getPageStartDefaultsForType, getRectFromProps, getSpringFromProps, Shar
 import { isNumber } from '../../utils/types';
 import { Screen } from '../../platform';
 import { CORE_ANIMATION_DEFAULTS } from '../../utils/common';
-import { iOSNativeHelper } from '../../utils/native-helper';
+import { ios as iOSUtils } from '../../utils/native-helper';
 
 interface PlatformTransitionInteractiveState extends TransitionInteractiveState {
 	transitionContext?: UIViewControllerContextTransitioning;
@@ -81,7 +81,7 @@ export class SharedTransitionHelper {
 								// in case the image is loaded async, we need to update the snapshot when it changes
 								// todo: remove listener on transition end
 								presentedView.on('imageSourceChange', () => {
-									snapshot.image = iOSNativeHelper.snapshotView(presentedSharedElement, Screen.mainScreen.scale);
+									snapshot.image = iOSUtils.snapshotView(presentedSharedElement, Screen.mainScreen.scale);
 									snapshot.tintColor = presentedSharedElement.tintColor;
 								});
 
@@ -89,7 +89,7 @@ export class SharedTransitionHelper {
 								snapshot.contentMode = presentedSharedElement.contentMode;
 							}
 
-							iOSNativeHelper.copyLayerProperties(snapshot, presentingSharedElement, pageEndProps?.propertiesToMatch);
+							iOSUtils.copyLayerProperties(snapshot, presentingSharedElement, pageEndProps?.propertiesToMatch as any);
 							snapshot.clipsToBounds = true;
 							// console.log('---> snapshot: ', snapshot);
 
@@ -97,7 +97,7 @@ export class SharedTransitionHelper {
 							const endFrame = presentedSharedElement.convertRectToView(presentedSharedElement.bounds, transitionContext.containerView);
 							snapshot.frame = startFrame;
 							if (SharedTransition.DEBUG) {
-								console.log('---> ', presentingView.sharedTransitionTag, ' frame:', iOSNativeHelper.printCGRect(snapshot.frame));
+								console.log('---> ', presentingView.sharedTransitionTag, ' frame:', iOSUtils.printCGRect(snapshot.frame));
 							}
 
 							transition.sharedElements.presenting.push({
@@ -151,11 +151,11 @@ export class SharedTransitionHelper {
 									await pageEndProps?.callback(independentView, 'present');
 								}
 
-								let snapshot: UIImageView;
+								// let snapshot: UIImageView;
 								// if (isPresented) {
 								// 	snapshot = UIImageView.alloc().init();
 								// } else {
-								snapshot = UIImageView.alloc().initWithImage(iOSNativeHelper.snapshotView(independentSharedElement, Screen.mainScreen.scale));
+								const snapshot = UIImageView.alloc().initWithImage(iOSUtils.snapshotView(independentSharedElement, Screen.mainScreen.scale));
 								// }
 
 								if (independentSharedElement instanceof UIImageView) {
@@ -184,7 +184,7 @@ export class SharedTransitionHelper {
 								snapshot.frame = startFrame; //startFrameAdjusted;
 								// }
 								if (SharedTransition.DEBUG) {
-									console.log('---> ', independentView.sharedTransitionTag, ' frame:', iOSNativeHelper.printCGRect(snapshot.frame));
+									console.log('---> ', independentView.sharedTransitionTag, ' frame:', iOSUtils.printCGRect(snapshot.frame));
 								}
 
 								const endFrameRect = getRectFromProps(pageEndProps);
@@ -289,18 +289,18 @@ export class SharedTransitionHelper {
 							presentingMatch.snapshot.frame = correctedEndFrame;
 
 							// apply view and layer properties to the snapshot view to match the source/presented view
-							iOSNativeHelper.copyLayerProperties(presentingMatch.snapshot, presented.view.ios, presented.propertiesToMatch);
+							iOSUtils.copyLayerProperties(presentingMatch.snapshot, presented.view.ios, presented.propertiesToMatch as any);
 							// create a snapshot of the presented view
-							presentingMatch.snapshot.image = iOSNativeHelper.snapshotView(presented.view.ios, Screen.mainScreen.scale);
+							presentingMatch.snapshot.image = iOSUtils.snapshotView(presented.view.ios, Screen.mainScreen.scale);
 							// apply correct alpha
 							presentingMatch.snapshot.alpha = presentingMatch.endOpacity;
 
 							if (SharedTransition.DEBUG) {
-								console.log(`---> ${presentingMatch.view.sharedTransitionTag} animate to: `, iOSNativeHelper.printCGRect(correctedEndFrame));
+								console.log(`---> ${presentingMatch.view.sharedTransitionTag} animate to: `, iOSUtils.printCGRect(correctedEndFrame));
 							}
 						}
 						for (const independent of transition.sharedElements.independent) {
-							let endFrame: CGRect = independent.endFrame;
+							const endFrame: CGRect = independent.endFrame;
 							// if (independent.isPresented) {
 							// 	const updatedEndFrame = independent.view.ios.convertRectToView(independent.view.ios.bounds, transitionContext.containerView);
 							// 	endFrame = CGRectMake(updatedEndFrame.origin.x, updatedEndFrame.origin.y, independent.endFrame.size.width, independent.endFrame.size.height);
@@ -313,7 +313,7 @@ export class SharedTransitionHelper {
 							independent.snapshot.alpha = independent.endOpacity;
 
 							if (SharedTransition.DEBUG) {
-								console.log(`---> ${independent.view.sharedTransitionTag} animate to: `, iOSNativeHelper.printCGRect(independent.endFrame));
+								console.log(`---> ${independent.view.sharedTransitionTag} animate to: `, iOSUtils.printCGRect(independent.endFrame));
 							}
 						}
 					};
@@ -332,7 +332,7 @@ export class SharedTransitionHelper {
 							}
 						);
 					} else {
-						iOSNativeHelper.animateWithSpring({
+						iOSUtils.animateWithSpring({
 							...getSpringFromProps(pageEnd?.spring),
 							animations: () => {
 								animateProperties();
@@ -409,7 +409,7 @@ export class SharedTransitionHelper {
 						}
 
 						// take a new snapshot
-						data.snapshot.image = iOSNativeHelper.snapshotView(view, Screen.mainScreen.scale);
+						data.snapshot.image = iOSUtils.snapshotView(view, Screen.mainScreen.scale);
 
 						// find the currently visible view with the same sharedTransitionTag
 						const fromView = transition.sharedElements.presented.find((p) => p.view.sharedTransitionTag === data.view.sharedTransitionTag)?.view;
@@ -458,12 +458,12 @@ export class SharedTransitionHelper {
 						transition.presented.view.frame = CGRectMake(endFrame.x, endFrame.y, endFrame.width, endFrame.height);
 
 						for (const presenting of transition.sharedElements.presenting) {
-							iOSNativeHelper.copyLayerProperties(presenting.snapshot, presenting.view.ios, presenting.propertiesToMatch);
+							iOSUtils.copyLayerProperties(presenting.snapshot, presenting.view.ios, presenting.propertiesToMatch as any);
 							presenting.snapshot.frame = presenting.startFrame;
 							presenting.snapshot.alpha = presenting.startOpacity;
 
 							if (SharedTransition.DEBUG) {
-								console.log(`---> ${presenting.view.sharedTransitionTag} animate to: `, iOSNativeHelper.printCGRect(presenting.snapshot.frame));
+								console.log(`---> ${presenting.view.sharedTransitionTag} animate to: `, iOSUtils.printCGRect(presenting.snapshot.frame));
 							}
 						}
 
@@ -476,7 +476,7 @@ export class SharedTransitionHelper {
 							}
 
 							if (SharedTransition.DEBUG) {
-								console.log(`---> ${independent.view.sharedTransitionTag} animate to: `, iOSNativeHelper.printCGRect(independent.snapshot.frame));
+								console.log(`---> ${independent.view.sharedTransitionTag} animate to: `, iOSUtils.printCGRect(independent.snapshot.frame));
 							}
 						}
 					};
@@ -495,7 +495,7 @@ export class SharedTransitionHelper {
 							}
 						);
 					} else {
-						iOSNativeHelper.animateWithSpring({
+						iOSUtils.animateWithSpring({
 							...getSpringFromProps(pageReturn?.spring),
 							animations: () => {
 								animateProperties();
@@ -542,7 +542,7 @@ export class SharedTransitionHelper {
 			interactiveState.propertyAnimator = UIViewPropertyAnimator.alloc().initWithDurationDampingRatioAnimations(1, 1, () => {
 				for (const p of state.instance.sharedElements.presenting) {
 					p.snapshot.frame = p.startFrame;
-					iOSNativeHelper.copyLayerProperties(p.snapshot, p.view.ios, p.propertiesToMatch);
+					iOSUtils.copyLayerProperties(p.snapshot, p.view.ios, p.propertiesToMatch as any);
 
 					p.snapshot.alpha = 1;
 				}
